split JavaScript into multiple files for better structure and maintainability
