+++ index.html
@@ -6,6 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Sticky notes app</title>
     <link rel="stylesheet" href="css/style.css">
+    <script type="module" src="js/index.js"></script>
 </head>
 
 <body>
@@ -22,7 +23,6 @@
         <button id="clearAll" class="buttons">Clear All</button>
     </div>
 
-    <script src="js/index.js"></script>
 </body>
 
 </html>
